==> VI_data_collection/5_correlations/3_compare_exports_1.0.Rmd <==
---
title: "Concordance Analysis"
author: "Andrew M. Demetriou"
---

Aim: the aim of this notebook is to examine the correlation in scores between the original, and subsequent exports of machines scores. 

```{r setup, include=FALSE}
library('here')            # file logistics
library('data.table')      # data manipulation
library('tidyverse')       # data manipulation
library('corrr')           # correlation analysis
library('patchwork')       # multiple plots 
library('stringr')         # string manipulation
library("foreach")         # iterate over two lists
library("purrr")           # iterate over two lists

values <- c("POWER", "ACHIEVEMENT", "HEDONISM",  
            "STIMULATION", "SELF", "UNIVERSALISM", 
            "BENEVOLENCE", "TRADITION",  
            "CONFORMITY", "SECURITY")

# read in item ids
path <- here("VI_data_collection", "_data", "primary_data", "mxm_ids.RDS")
mxm_IDs <- readRDS(path)

# read in function
read_files <- function(file){fread(here(path, file))}

rm(path)
```

```{r}
# format dfs
format_dfs <- function(df){
  # make all column names uppercase
  colnames(df) <- toupper(colnames(df))
  
  # rename variable to conform to convention
  df <- df %>% rename(SELF = `SELF-DIRECTION`) %>% select(values)
  
  df$item_ID <- as.character(mxm_IDs$item_ID)
  
  return(df)
}
```

```{r}
# load and format og machines df
load(file = here("VI_data_collection", "_data","intermediary_data", "machines_as_fixed_mean_dfs.RDS"))

og_scores_df <- machines_as_fixed_mean_dfs[[1]] 

colnames(og_scores_df) <- gsub("glove840B300d_uniform_weight_lyrics", 
                               "glove_uniform", colnames(og_scores_df))
colnames(og_scores_df) <- gsub("glove840B300d_idf_weight_lyrics", 
                               "glove_idf", colnames(og_scores_df))
colnames(og_scores_df) <- gsub("googlenews_uniform_weight_lyrics", 
                               "google_uniform", colnames(og_scores_df))
colnames(og_scores_df) <- gsub("googlenews_idf_weight_lyrics", 
                               "google_idf", colnames(og_scores_df))

rm(machines_as_fixed_mean_dfs)
```

```{r}
correlate_values <- function(df){
  
  cor_list <- lapply(values, function(value){
    value_2 <- paste0(value, "_")
  
    cor(df[[value]], df[[value_2]])
  })

  names(cor_list) <- values

  results_df <- data.frame(value = rep(names(cor_list), sapply(cor_list, length)),
                 correlation = unlist(cor_list))
  rownames(results_df) <- NULL
  
  return(results_df)
}
```

```{r}
make_results_df <-function(
    machine_scores, 
    setups, 
    format = "list",
    values = c("POWER", "ACHIEVEMENT", "HEDONISM",  
            "STIMULATION", "SELF", "UNIVERSALISM", 
            "BENEVOLENCE", "TRADITION",  
            "CONFORMITY", "SECURITY")){
  
  results_list <- lapply(as.list(setups), function(setup){
  
  # subset export
  #  setup <- setups[[1]]
  # machine_scores <- machine_scores_dfs  

    if(format == "list"){
      new_scores_df <- machine_scores[[setup]]
    }else if(format == "dataframe"){
      new_scores_df <- machine_scores %>%
        select(item_ID, contains(setup))
    }else(print("enter either 'dataframe' or 'list' as format"))
  
  # subset original scores
  old_scores_df <- og_scores_df %>% 
    select(item_ID, contains(setup))
  
  # rename columns to conform to names of new scores df
  colnames(old_scores_df) <- gsub(setup, "", colnames(old_scores_df))
  colnames(new_scores_df) <- gsub(paste0("_", setup), "", colnames(new_scores_df))
  
  # merge two dfs
  df <- merge(old_scores_df, new_scores_df, by="item_ID")
  # estimate correlations
  results_dt <- correlate_values(df)
  # name result after setup
  results_dt$model <- setup
  # convert to data table
  setDT(results_dt)
  
  return(results_dt)
})

# bind list of dfs into a single list
results_df <- rbindlist(results_list)
return(results_df)

}
```

```{r}
plot_cor_results <- function(results_df){
  results_df %>%
  ggplot(., aes(x=model, y=value, fill = correlation)) +
    geom_tile(color = "white") + 
    geom_text(aes(label = round(correlation, 2)), color = "white")+
  scale_fill_gradient(low = "blue", high = "red") +
  theme(axis.text.x = element_text(angle = 90))
}
```

```{r}
# third export

# path to files
path <- here("VI_data_collection", "_data", "machine_scores", "third_export")
files <- list.files(path)

# read in files
machine_scores_dfs <- lapply(files, read_files)

# name files
names(machine_scores_dfs) <- gsub(".csv", "", files)

# format downloaded machine scores
machine_scores_dfs <- lapply(machine_scores_dfs, format_dfs)

rm(path, files)
```

```{r}
grid_1 <- c("google", "glove")
grid_2 <- c("idf", "uniform")

grid <- expand.grid(grid_1, grid_2)
colnames(grid) <- c("machines", "weight")

setups <- paste0(grid$machines, "_", grid$weight)

rm(grid_1, grid_2, grid)
```

```{r}
# third export results
results_df <- make_results_df(machine_scores_dfs, setups)
results_df %>% plot_cor_results()
```


```{r}
# fourth export

# path to files
path <- here("VI_data_collection", "_data", "machine_scores", "fourth_export")
files <- list.files(path)

# read in function
machine_scores_dfs <- lapply(files, read_files)

# name files
names(machine_scores_dfs) <- gsub(".csv", "", files)

# cut out word counting
machine_scores_dfs <-  machine_scores_dfs[40:43]

rm(path, files)
```

```{r}
# format downloaded machine scores
machine_scores_dfs <- lapply(machine_scores_dfs, format_dfs)

names(machine_scores_dfs) <- c("glove_idf",  "glove_uniform",  "google_idf", "google_uniform")
```

```{r}
# fourth export, pretrained model results
results_df <- make_results_df(machine_scores_dfs, setups)
results_df %>% plot_cor_results()
```

```{r}
cols <- colnames(og_scores_df)
cols <- gsub("_mean", "", cols)
cols <- gsub("_uniform", "_uni", cols)
cols <- gsub("_split", "_cv", cols)

colnames(og_scores_df) <- cols

rm(cols)
```

```{r}
# fourth export

# path to files
path <- here("VI_data_collection", "_data", "machine_scores", "fourth_export")
files <- list.files(path)

# read in function
machine_scores_dfs <- lapply(files, read_files)

# name files
names(machine_scores_dfs) <- gsub(".csv", "", files)

# cut out word counting
machine_scores_dfs <-  machine_scores_dfs[1:43]

rm(path, files)
```

```{r}
# format downloaded machine scores
machine_scores_dfs <- lapply(machine_scores_dfs, format_dfs)
```

```{r}
estimate_average_score <- function(df){
    
 # estimate average of scores for each machine setup 
 average_scores_df <- lapply(values, function(x){
   # select all columns that start with the value
   df <- df %>% select(all_of(starts_with(x)))
   # compute row means
   scores <- df %>% mutate(average = rowMeans(.)) %>% select(average)
   # rename column
   colnames(scores) <- x
   return(scores)
  })

 # rename items in the list
 names(average_scores_df) <- values
 
 # coerce list to a dataframe
 average_scores_df <- as.data.frame(do.call(cbind, average_scores_df))
 
 return(average_scores_df)
}
```

```{r}
make_mean_df_list <- function(machines_as_fixed_df){
  
  # create a df of machines and weights to iterate over
  machines <- data.frame(machine = c("far", "far", "cv", "cv"), 
                         weight = c("idf", "uni", "idf", "uni"))

  # iterate over all combinations in machines df
  foreach(i = machines$machine, j = machines$weight) %do% {
    #subset main dataset by combination
    df <- machines_as_fixed_df %>% select(intersect(contains(i), contains(j)))
    
  
    # compute single average variable for each machine setup
    average_scores_df <- estimate_average_score(df)
    average_scores_df$model  <- i
    average_scores_df$weight <- j
    average_scores_df$method <- "mean"
    average_scores_df$item_ID <- machines_as_fixed_df$item_ID
    base::assign(paste0(i,"_",j, "_average"), average_scores_df)
  }

  # output list for each setup  
  df_list <- list(far_idf_average = far_idf_average, 
                  far_uni_average = far_uni_average, 
                  cv_idf_average = cv_idf_average, 
                  cv_uni_average = cv_uni_average)
  
  output_df <- rbindlist(df_list) %>%
      pivot_wider(., id_cols = item_ID, 
            names_from = c(model, weight, method), 
            values_from = values)
  
}
#s<-make_mean_df_list(machines_as_fixed_df_list[[3]])
```

```{r}
machine_scores_df <- rbindlist(machine_scores_dfs, idcol=T)
machine_scores_df <- machine_scores_df %>% pivot_wider(names_from = .id, values_from = values)
machine_scores_df <- make_mean_df_list(machine_scores_df)
colnames(machine_scores_df) <- gsub("_mean", "", colnames(machine_scores_df)) 
```

```{r}
#grid_1 <- c("google", "glove", "cv", "far")
grid_1 <- c("cv", "far")
grid_2 <- c("idf", "uni")

grid <- expand.grid(grid_1, grid_2)
colnames(grid) <- c("machines", "weight")

setups <- paste0(grid$machines, "_", grid$weight)

rm(grid_1, grid_2, grid)
```

```{r}
results_df <- make_results_df(machine_scores = machine_scores_df, format = "dataframe", setups)
results_df %>% plot_cor_results()
```

```{r}
machines_as_subjects_df <- readRDS(here("VI_data_collection", "_data", "machine_response_data", "machines_as_subjects_df.RDS"))



df <- machines_as_subjects_df[[1]]

df <- df %>% filter(participant_ID == "mxm_faruqui_0_uniform_weight_lyrics")

df_2 <- machine_scores_dfs[["far_0_idf"]]

df_3 <- merge(df, df_2, by = "item_ID")

cor(df_3$POWER.x, df_3$POWER.y)
cor(df_3$ACHIEVEMENT.x, df_3$ACHIEVEMENT.y)
cor(df_3$HEDONISM.x, df_3$HEDONISM.y)
```
